app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pymysql>=1.1
pillow>=10.0
